% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_hms_results}
\alias{cc_hms_results}
\title{Cherry Creek Dam HEC-HMS Routing Results}
\format{
A data frame with 5 columns:
\describe{
\item{time_hr}{Time in hours (integer)}
\item{inflow_cfs}{Inflow in cubic-feet per second (CFS)}
\item{elevation_ft}{Routed reservoir stage/elevation data in FT-NAVD88}
\item{stor_acft}{Routed reservoir storage (ACRE-FT) corresponding to reservoir elevation (elev_ft).}
\item{outflow_cfs}{Routed discharge/outflow (CFS) corresponding to reservoir elevation (elev_ft).}
}
}
\usage{
cc_hms_results
}
\description{
Routing results from HEC-HMS using the Cherry Creek reservoir model/geometry, inflow hydrograph, and initial reservoir elevation.
}
\keyword{datasets}
