% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_inflowhydro}
\alias{cc_inflowhydro}
\title{Cherry Creek Inflow Hydrograph}
\format{
A data frame with 2 columns:
\describe{
\item{time_hr}{Time in hours (integer)}
\item{inflow_cfs}{Inflow in cubic-feet per second (CFS)}
}
}
\usage{
cc_inflowhydro
}
\description{
Inflow hydrograph for an example flood event at Cherry Creek Dam.
}
\keyword{datasets}
