% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc_init_elev}
\alias{cc_init_elev}
\title{Cherry Creek Dam Initial Elevation}
\format{
A single numeric value representing the initial reservoir elevation (FT-NAVD88).
}
\usage{
cc_init_elev
}
\description{
Initial/starting elevation (FT-NAVD88) for Cherry Creek Dam reservoir routing.
This elevation was used in HEC-HMS.
}
\keyword{datasets}
