% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bins.R
\name{create_bins}
\alias{create_bins}
\title{Create Stratified Bins in EVI (Gumbel) Space}
\usage{
create_bins(min_aep, max_aep, n_bins)
}
\arguments{
\item{min_aep}{Minimum annual exceedance probability (e.g., 1e-8)}

\item{max_aep}{Maximum annual exceedance probability (e.g., 0.99)}

\item{n_bins}{Number of bins (e.g., 50)}
}
\value{
Vector of bin boundaries (length n_bins + 1)
}
\description{
Create Stratified Bins in EVI (Gumbel) Space
}
