% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_bf_parameter_sets}
\alias{jmd_bf_parameter_sets}
\title{John Martin Dam BestFit Parameter Sets}
\format{
A data frame with 10000 row and 3 columns:
\describe{
\item{mean_log}{Mean of log-transformed values}
\item{sd_log}{Standard deviation of log-transformed values}
\item{skew_log}{Skewness of log-transformed values}
}
}
\usage{
jmd_bf_parameter_sets
}
\description{
LP3 distribution parameters sets of volume-frequency curve results from
RMC-BestFit 2.0.
}
\keyword{datasets}
