% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_empirical_stage_por}
\alias{jmd_empirical_stage_por}
\title{John Martin Dam Empirical Stage-Frequency (Full POR)}
\format{
A data frame with 82 rows and 3 columns:
\describe{
\item{wy}{Water year}
\item{stage_ft}{Stage (FT-NAVD88)}
\item{plot_posit}{Plotting position}
}
}
\usage{
jmd_empirical_stage_por
}
\description{
Empirical stage-frequency results for John Martin Dam full period of record.
}
\keyword{datasets}
