% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_empirical_stage_wy1980}
\alias{jmd_empirical_stage_wy1980}
\title{John Martin Dam Empirical Stage-Frequency (WY 1980-2024)}
\format{
A data frame with 45 rows and 3 columns:
\describe{
\item{wy}{Water year}
\item{stage_ft}{Stage (FT-NAVD88)}
\item{plot_posit}{Plotting position}
}
}
\usage{
jmd_empirical_stage_wy1980
}
\description{
Empirical stage-frequency results for John Martin Dam, water years 1980-2024.
}
\keyword{datasets}
