% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_empirical_stage_wy1980_pt}
\alias{jmd_empirical_stage_wy1980_pt}
\title{John Martin Dam Empirical Stage-Frequency with Perception Threshold}
\format{
A data frame with 54 rows and 3 columns:
\describe{
\item{wy}{Water year}
\item{stage_ft}{Stage (FT-NAVD88)}
\item{plot_posit}{Plotting position}
}
}
\usage{
jmd_empirical_stage_wy1980_pt
}
\description{
Empirical stage-frequency results for John Martin Dam (WY 1980-2024) with
perception threshold applied.
}
\keyword{datasets}
