% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_jun1965}
\alias{jmd_hydro_jun1965}
\title{John Martin Dam June 1965 Hydrograph}
\format{
A data frame with 241 rows and 2 columns:
\describe{
\item{time_hr}{Time in hours}
\item{inflow_cfs}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_jun1965
}
\description{
Inflow hydrograph shape from June 1965 flood event (from HEC-HMS).
}
\keyword{datasets}
