% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_jun1965_15min}
\alias{jmd_hydro_jun1965_15min}
\title{JMD June 1965 Hydrograph (15 minute intervals)}
\format{
A data frame with 481 rows and 4 columns:
\describe{
\item{Ordinate}{Timeseries Ordinate}
\item{Date}{Date mm/dd/yyyy}
\item{Time}{Time in 00:00}
\item{Flow}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_jun1965_15min
}
\description{
Inflow hydrograph shape from June 1965 flood event (from HEC-HMS).
}
\examples{
sapply(jmd_hydro_jun1965_15min,class)
head(jmd_hydro_jun1965_15min)
plot(jmd_hydro_jun1965_15min$Ordinate/4, jmd_hydro_jun1965_15min$Flow,
     xlab = "Hour", ylab = "Inflow (cfs)",
     type = "l")
}
\keyword{datasets}
