% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_may1955}
\alias{jmd_hydro_may1955}
\title{John Martin Dam May 1955 Hydrograph}
\format{
A data frame with 121 rows and 2 columns:
\describe{
\item{time_hr}{Time in hours}
\item{inflow_cfs}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_may1955
}
\description{
Inflow hydrograph shape from May 1955 flood event.
}
\keyword{datasets}
