% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_pmf}
\alias{jmd_hydro_pmf}
\title{JMD PMF Hydrograph}
\format{
A data frame with 192 rows and 2 columns:
\describe{
\item{Ordinate}{Timeseries Ordinate}
\item{Date}{Date mm/dd/yyyy}
\item{Time}{Time in 00:00}
\item{Flow}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_pmf
}
\description{
Inflow hydrograph shape for Probable Maximum Flood (PMF).
}
\examples{
head(jmd_hydro_pmf)
\donttest{
plot(jmd_hydro_pmf$Ordinate, jmd_hydro_pmf$Flow,
     xlab = "Ordinate", ylab = "Inflow (cfs)",
     type = "l")
}
}
\keyword{datasets}
