% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_inflowhydro}
\alias{jmd_inflowhydro}
\title{John Martin Dam Inflow Hydrograph}
\format{
A data frame with 2 columns:
\describe{
\item{time_hr}{Time in hours (integer)}
\item{inflow_cfs}{Inflow in cubic-feet per second (CFS)}
}
}
\usage{
jmd_inflowhydro
}
\description{
Inflow hydrograph for hypothetical PMF flood event at John Martin Dam.
}
\keyword{datasets}
