% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_init_elev}
\alias{jmd_init_elev}
\title{John Martin Dam Initial Elevation}
\format{
A single numeric value representing the initial reservoir elevation (FT-NAVD88).
}
\usage{
jmd_init_elev
}
\description{
Initial/starting elevation (FT-NAVD88) for JMD reservoir routing.
This elevation was used in HEC-HMS.
}
\keyword{datasets}
