% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_por_inflow}
\alias{jmd_por_inflow}
\title{JMD Period of Record Inflow}
\format{
A data frame with 40,908 rows and 4 columns:
\describe{
\item{timestep}{Timestep index}
\item{date}{Date}
\item{time}{Time}
\item{flow_cfs}{Daily average inflow (cfs)}
}
}
\usage{
jmd_por_inflow
}
\description{
Daily inflow at JMD for the available period of record.
}
\examples{
sapply(jmd_por_inflow,class)
head(jmd_por_inflow)
plot(as.Date(jmd_por_inflow$date,tryFormats = c("\%m/\%d/\%Y")), jmd_por_inflow$flow_cfs,
      xlab = "Year", ylab = "Inflow (cfs)",
      type = "l")
}
\keyword{datasets}
