% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_por_stage}
\alias{jmd_por_stage}
\title{John Martin Dam Period of Record Stage}
\format{
A data frame with 29,858 rows and 4 columns:
\describe{
\item{timestep}{Timestep index}
\item{date}{Date}
\item{time}{Time}
\item{stage_ft}{Reservoir stage (FT-NAVD88)}
}
}
\usage{
jmd_por_stage
}
\description{
Daily stage data from the full John Martin Dam period of record (1943-2024).
}
\keyword{datasets}
