% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_resmodel}
\alias{jmd_resmodel}
\title{JMD Reservoir Model}
\format{
A data frame with 3 columns:
\describe{
\item{stage_ft}{Reservoir stage/elevation data (FT-NAVD88)}
\item{stor_acft}{Cumulative reservoir storage (ACRE-FT) corresponding to reservoir elevation (elev_ft).}
\item{discharge_cfs}{Cumulative discharge/outflow (CFS) corresponding to reservoir elevation (elev_ft).}
}
}
\usage{
jmd_resmodel
}
\description{
Elevation-storage-discharge relationship for John McGraw Reservoir. Reservoir models must be monotonic.
}
\examples{
head(jmd_resmodel)
\donttest{
plot(jmd_resmodel[[1]], jmd_resmodel[[2]],
     xlab = "Elevation (ft)", ylab = "Storage (acre-ft)",
     type = "l")
}
}
\keyword{datasets}
