% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_rfa_expected}
\alias{jmd_rfa_expected}
\title{John Martin Dam RFA Results - Expected Only}
\format{
A data frame with 10000 row and 3 columns:
\describe{
\item{AEP}{Annual Exceedace Probabilities}
\item{Expected}{Expected stages corresponding to AEPs}
}
}
\usage{
jmd_rfa_expected
}
\description{
John Martin Dam, expected only stage-frequency results from RFA.
}
\keyword{datasets}
