% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_rfa_full}
\alias{jmd_rfa_full}
\title{Jay McGraw Dam RFA Results - Full Uncertainty}
\format{
A data frame with 10000 rows and 3 columns:
\describe{
\item{AEP}{Annual Exceedance Probabilities}
\item{Upper}{Upper 95\% CI of realizations}
\item{Lower}{Lower 5\% CI of realizations}
\item{Expected}{Expected stages corresponding to AEPs}
\item{Median}{Median stages corresponding to AEPs}
}
}
\usage{
jmd_rfa_full
}
\description{
Jay McGraw Dam, full uncertainty stage-frequency results from RFA.
}
\examples{
head(jmd_rfa_full)
}
\keyword{datasets}
