% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_seasonality}
\alias{jmd_seasonality}
\title{John Martin Dam Flood Seasonality}
\format{
A data frame with 12 rows and 4 columns:
\describe{
\item{month}{Month (1-12)}
\item{frequency}{Event frequency}
\item{relative_frequency}{Relative frequency}
\item{cume_rel_frequency}{Cumulative relative frequency}
}
}
\usage{
jmd_seasonality
}
\description{
Results of flood seasonality analysis from the full John Martin Dam period
of record. Determined using: threshold flow of 5,000 cfs, critical duration
of 2 days, max events per year of 5, minimum days between events of 7.
}
\keyword{datasets}
