% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_stage_duration}
\alias{jmd_stage_duration}
\title{John Martin Dam Monthly Stage Duration Curves}
\format{
A data frame with 45 rows and 13 columns:
\describe{
\item{Probability}{Probability of exceedance}
\item{January}{Stage for January (FT-NAVD88)}
\item{February}{Stage for February (FT-NAVD88)}
\item{March}{Stage for March (FT-NAVD88)}
\item{April}{Stage for April (FT-NAVD88)}
\item{May}{Stage for May (FT-NAVD88)}
\item{June}{Stage for June (FT-NAVD88)}
\item{July}{Stage for July (FT-NAVD88)}
\item{August}{Stage for August (FT-NAVD88)}
\item{September}{Stage for September (FT-NAVD88)}
\item{October}{Stage for October (FT-NAVD88)}
\item{November}{Stage for November (FT-NAVD88)}
\item{December}{Stage for December (FT-NAVD88)}
}
}
\usage{
jmd_stage_duration
}
\description{
Monthly stage duration curves from Water Year 1980-2024.
}
\keyword{datasets}
