% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_vfc}
\alias{jmd_vfc}
\title{John Martin Dam Volume-Frequency Curve}
\format{
A data frame with 25 rows and 5 columns:
\describe{
\item{aep}{Annual exceedance probability}
\item{ci_95}{95\% credible interval}
\item{ci_5}{5\% credible interval}
\item{posterior_predictive}{Posterior predictive value}
\item{posterior_mode}{Posterior mode value}
}
}
\usage{
jmd_vfc
}
\description{
Tabular volume-frequency curve results from RMC-BestFit 2.0.
}
\keyword{datasets}
