% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_vfc_parameters}
\alias{jmd_vfc_parameters}
\title{John Martin Dam VFC Parameters}
\format{
A data frame with 1 row and 5 columns:
\describe{
\item{mean_log}{Mean of log-transformed values}
\item{sd_log}{Standard deviation of log-transformed values}
\item{skew_log}{Skewness of log-transformed values}
\item{erl}{Estimated effective record length}
\item{duration}{Inflow-volume duration in days}
}
}
\usage{
jmd_vfc_parameters
}
\description{
LP3 distribution parameters of volume-frequency curve results from
RMC-BestFit 2.0.
}
\keyword{datasets}
