% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_wy1980_stage}
\alias{jmd_wy1980_stage}
\title{JMD Stage (WY 1980-2024)}
\format{
A data frame with 29,858 rows and 4 columns:
\describe{
\item{timestep}{Timestep index}
\item{date}{Date}
\item{time}{Time}
\item{stage_ft}{Reservoir stage (FT-NAVD88)}
}
}
\usage{
jmd_wy1980_stage
}
\description{
Daily stage data from Water Year 1980-2024.
}
\examples{
sapply(jmd_wy1980_stage,class)
head(jmd_wy1980_stage)
plot(as.Date(jmd_wy1980_stage$date,tryFormats = c("\%m/\%d/\%Y")), jmd_wy1980_stage$stage_ft,
      xlab = "Year", ylab = "Inflow (cfs)",
      type = "l")
}
\keyword{datasets}
