% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jst_inflowhydro}
\alias{jst_inflowhydro}
\title{JST Inflow Hydrograph (Routing Validation)}
\format{
Data frame
}
\usage{
jst_inflowhydro
}
\description{
JST Inflow Hydrograph (Routing Validation)
}
\keyword{datasets}
