% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jst_init_elev}
\alias{jst_init_elev}
\title{JST Initial Elevation (Routing Validation)}
\format{
Numeric
}
\usage{
jst_init_elev
}
\description{
JST Initial Elevation (Routing Validation)
}
\keyword{datasets}
