% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{power_function}
\alias{power_function}
\title{Power Function}
\usage{
power_function(x)
}
\arguments{
\item{x}{Highest power of 10 to us}
}
\value{
Vector of 10 raised sequentially from the 1st to the nth power.
}
\description{
Raises 10 to the 1st through the nth power
}
\keyword{internal}
