% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson3.R
\name{pp3}
\alias{pp3}
\title{Pearson Type III Cumulative Distribution Function}
\usage{
pp3(x, mu, sigma, gamma)
}
\arguments{
\item{x}{Vector of quantiles (in log10 space for Log-Pearson III)}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{gamma}{Skewness coefficient}
}
\value{
Vector of probabilities
}
\description{
Distribution function for the Pearson Type III distribution.
}
\examples{
# Single value
pp3(4.5, mu = 4.232, sigma = 0.153, gamma = 0.401)

# Vector of values
x = seq(3.5, 5.0, length.out = 100)
p = pp3(x, mu = 4.232, sigma = 0.153, gamma = 0.401)
plot(x, p, type = "l")
}
