% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp3.R
\name{qp3}
\alias{qp3}
\title{Pearson Type III Inverse CDF (Quantile Function)}
\usage{
qp3(p, mu, sigma, gamma)
}
\arguments{
\item{p}{Vector of probabilities (between 0 and 1).}

\item{mu}{Mean of the distribution.}

\item{sigma}{Standard deviation of the distribution.}

\item{gamma}{Skewness coefficient of the distribution.}
}
\value{
Vector of quantiles corresponding to the input probabilities.
}
\description{
Computes quantiles from a Pearson Type III distribution given
probabilities and distribution parameters (mean, standard deviation, skewness).
}
\details{
NOTE: This is the non-vectorized version of the function (it will be used in a loop)
Vectorized version is part of future development.

The Pearson Type III distribution is parameterized by mean (\code{mu}),
standard deviation (\code{sigma}), and skewness (\code{gamma}). These are converted
internally to location, scale, and shape parameters.

When skewness is near zero (\code{abs(gamma) < 1E-3}), the normal distribution
is used as an approximation.
}
\examples{
# Single quantile
qp3(0.99, mu = 10, sigma = 2, gamma = 0.5)

# Multiple quantiles
qp3(c(0.5, 0.9, 0.99), mu = 10, sigma = 2, gamma = 0.5)

# Use with JMD volume-frequency parameters
qp3(0.99,
    mu = jmd_vfc_parameters$mean_log,
    sigma = jmd_vfc_parameters$sd_log,
    gamma = jmd_vfc_parameters$skew_log)
}
