% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_vfc.R
\name{bootstrap_vfc}
\alias{bootstrap_vfc}
\title{Bootstrap option for VFC (RFA-style w/ ERL)}
\usage{
bootstrap_vfc(bestfit_postmode, dist = "LP3", ERL = 150, Nboots = 10000)
}
\arguments{
\item{bestfit_postmode}{Vector of distribution parameters from RMC-BestFit.
For LP3: columns are mean (log), sd (log), skew (log).
For GEV: columns are location, scale, shape.}

\item{dist}{Distribution type. Either \code{"LP3"} (default) or \code{"GEV"}.}

\item{ERL}{psuedo effective record length for bootstrapping}

\item{Nboots}{number of bootstraps}
}
\value{
A list containing:
\describe{
\item{params}{Matrix of bootstrapped parameters}
\item{dist}{Distribution selected (LP3 or GEV)}
\item{postmode}{Posterior mode parameters provided to bootstrap}
\item{ERL}{pseudo effective record length provided to bootstrap}
}
}
\description{
Generates matrix of parameters using posterior mode/mean and ERL
}
\examples{
# Sample using JMD VFC parameters
# samples <- flow_frequency_sampler(jmd_vfc_parameters)
}
\seealso{
\code{\link[=flow_frequency_sampler]{flow_frequency_sampler()}}
}
