% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_apr1999}
\alias{jmd_hydro_apr1999}
\title{John Martin Dam April 1999 Hydrograph}
\format{
A data frame with 241 rows and 4 columns:
\describe{
\item{Ordinate}{Timeseries Ordinate}
\item{Date}{Date mm/dd/yyyy}
\item{Time}{Time in 00:00}
\item{Flow}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_apr1999
}
\description{
Inflow hydrograph shape from April 1999 flood event.
}
\keyword{datasets}
