% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_jun1921}
\alias{jmd_hydro_jun1921}
\title{John Martin Dam June 1921 Hydrograph}
\format{
A data frame with 169 rows and 4 columns:
\describe{
\item{Ordinate}{Timeseries Ordinate}
\item{Date}{Date mm/dd/yyyy}
\item{Time}{Time in 00:00}
\item{Flow}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_jun1921
}
\description{
Inflow hydrograph shape from June 1921 flood event.
}
\keyword{datasets}
