% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_pmf}
\alias{jmd_hydro_pmf}
\title{John Martin Dam PMF Hydrograph}
\format{
A data frame with 192 rows and 2 columns:
\describe{
\item{time_hr}{Time in hours}
\item{inflow_cfs}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_pmf
}
\description{
Inflow hydrograph shape for Probable Maximum Flood (PMF).
}
\keyword{datasets}
