% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_hydro_sdf}
\alias{jmd_hydro_sdf}
\title{John Martin Dam SDF Hydrograph}
\format{
A data frame with 337 rows and 2 columns:
\describe{
\item{time_hr}{Time in hours}
\item{inflow_cfs}{Inflow (cfs)}
}
}
\usage{
jmd_hydro_sdf
}
\description{
Inflow hydrograph shape for Spillway Design Flood (SDF).
}
\keyword{datasets}
