% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_por_inflow}
\alias{jmd_por_inflow}
\title{John Martin Dam Period of Record Inflow}
\format{
A data frame with 40,908 rows and 4 columns:
\describe{
\item{timestep}{Timestep index}
\item{date}{Date}
\item{time}{Time}
\item{flow_cfs}{Daily average inflow (cfs)}
}
}
\usage{
jmd_por_inflow
}
\description{
Daily inflow at John Martin Dam for the available period of record (1943-2024).
}
\keyword{datasets}
