% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_resmodel}
\alias{jmd_resmodel}
\title{John Martin Dam Reservoir Model}
\format{
A data frame with 3 columns:
\describe{
\item{elev_ft}{Reservoir stage/elevation data (FT-NAVD88)}
\item{stor_acft}{Cumulative reservoir storage (ACRE-FT) corresponding to reservoir elevation (elev_ft).}
\item{outflow_cfs}{Cumulative discharge/outflow (CFS) corresponding to reservoir elevation (elev_ft).}
}
}
\usage{
jmd_resmodel
}
\description{
Elevation-storage-discharge relationship for John Martin Reservoir (Arkansas River). Reservoir models must be monotonic.
}
\keyword{datasets}
