% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jmd_rfa_full}
\alias{jmd_rfa_full}
\title{John Martin Dam RFA Results - Full Uncertainty}
\format{
A data frame with 10000 row and 3 columns:
\describe{
\item{AEP}{Annual Exceedace Probabilities}
\item{Upper}{Annual Exceedace Probabilities}
\item{Lower}{Annual Exceedace Probabilities}
\item{Expected}{Expected stages corresponding to AEPs}
\item{Median}{Median stages corresponding to AEPs}
}
}
\usage{
jmd_rfa_full
}
\description{
John Martin Dam, expected only stage-frequency results from RFA.
}
\keyword{datasets}
