% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson3.R
\name{qp3}
\alias{qp3}
\title{Pearson Type III Quantile Function}
\usage{
qp3(p, mu, sigma, gamma)
}
\arguments{
\item{p}{Vector of probabilities}

\item{mu}{Mean}

\item{sigma}{Standard deviation}

\item{gamma}{Skewness coefficient}
}
\value{
Vector of quantiles (in log10 space for Log-Pearson III)
}
\description{
Quantile function (inverse CDF) for the Pearson Type III distribution.
}
\examples{
# Single value
qp3(0.5, mu = 4.232, sigma = 0.153, gamma = 0.401)

# Vector of values
p = seq(0.01, 0.99, length.out = 100)
q = qp3(p, mu = 4.232, sigma = 0.153, gamma = 0.401)
plot(p, q, type = "l")
}
