% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hydrograph.R
\name{scale_hydrograph}
\alias{scale_hydrograph}
\title{Scale Hydrograph}
\usage{
scale_hydrograph(hydrograph_shape, observed_volume, sampled_volume)
}
\arguments{
\item{hydrograph_shape}{Data frame with two columns: time (hours) and
inflow (cfs). Must be in that order.}

\item{observed_volume}{Maximum n-day inflow volume from input hydrograph shape.}

\item{sampled_volume}{Sampled inflow volume from volume-frequency curve.}
}
\value{
A data frame with two columns: \code{time_hrs} and \code{inflow_cfs}.
}
\description{
Scales inflow hydrograph given a sampled inflow volume from the volume-frequency
curve. Scale factor is defined by the sampled inflow volume and the maximum
volume of the corresponding duration.
}
\examples{
# Scale example hydrograph
scaled <- scale_hydrograph(cc_inflowhydro, observed_vol = 2500, sampled_volume = 5000,
                           critical_duration = 2, routing_days = 10)
}
