% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_hydrograph.R
\name{scale_hydrograph}
\alias{scale_hydrograph}
\title{Scale Hydrograph}
\usage{
scale_hydrograph(
  hydrograph_shape,
  sampled_volume,
  critical_duration,
  routing_days
)
}
\arguments{
\item{hydrograph_shape}{Data frame with two columns: time (hours) and
inflow (cfs). Must be in that order.}

\item{sampled_volume}{Sampled inflow volume from volume-frequency curve.}

\item{critical_duration}{Critical duration in days.}

\item{routing_days}{Desired length of routing simulation in days.}
}
\value{
A data frame with two columns: \code{time_hrs} and \code{inflow_cfs}.
}
\description{
Scales inflow hydrograph given a sampled inflow volume from the volume-frequency
curve. Scale factor is defined by the sampled inflow volume and the maximum
volume of the corresponding duration.
}
\examples{
# Scale example hydrograph
scaled <- scale_hydrograph(cc_inflowhydro, sampled_volume = 5000,
                           critical_duration = 2, routing_days = 10)
}
