% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{theme_rfar_conceptual}
\alias{theme_rfar_conceptual}
\title{rfaR ggplot theme for conceptual example}
\usage{
theme_rfar_conceptual()
}
\description{
custom ggplot2 theme for rfaR-Realization-Conceptual.Rmd
}
